refactor: フルスクリーン画像を component に移しナビはページ側に記述
- HistoryFullscreenImageViewer を presentation/component に追加
- util の手続きヘルパーを削除し各ページで Navigator.push を使用

Refs: #75
